TensorBoard: Improved Plugin Extensibility (#90)
This RFC proposes a friction-free approach to enabling TensorBoard
contributors to create and share TensorBoard plugins as installable
extensions.

TensorBoard is composed of various plugins that show up as separate
dashboards to users (e.g., scalars, graphs, hparams, histograms, etc.).
Today. anyone can contribute a new plugin by forking TensorBoard, adding
their plugin, and getting the PR merged. However, this can require a
lengthy review to make sure it is architected correctly and is aligned
with TensorBoard’s goals.

This RFC covers two changes that will make this process easier: Plugins
can be distributed as Python packages. Users can install them as
extensions to TensorBoard.  Plugins are no longer limited to use Polymer
but can instead use any frontend library or framework (e.g., React)

The intention is to enable any prospective contributor to easily be able
to create or reuse a visualization with tools familiar to them, and
enable users to try it out quickly. Plugins that are useful to many
users should still be added to TensorBoard, but this removes the
friction needed to start getting feedback or trying out something new.